+++ src/design/plone/contenttypes/adapters/searchabletext_indexers.py
@@ -2,11 +2,13 @@
 from collective.dexteritytextindexer.converters import (
     DefaultDexterityTextIndexFieldConverter,
 )
-from collective.dexteritytextindexer.interfaces import IDexterityTextIndexFieldConverter
-from z3c.relationfield.interfaces import IRelationChoice
-from z3c.relationfield.interfaces import IRelationList
+from collective.dexteritytextindexer.interfaces import (
+    IDexterityTextIndexFieldConverter,
+)
 from plone.dexterity.interfaces import IDexterityContent
 from z3c.form.interfaces import IWidget
+from z3c.relationfield.interfaces import IRelationChoice
+from z3c.relationfield.interfaces import IRelationList
 from zope.component import adapter
 from zope.interface import implementer
 
@@ -31,4 +33,6 @@ class RelationListFieldConverter(DefaultDexterityTextIndexFieldConverter):
         relations = self.field.get(self.context)
         if not relations:
             return ""
-        return " ".join([x.to_object.Title() for x in relations if x.to_object])
+        return " ".join(
+            [x.to_object.Title() for x in relations if x.to_object]
+        )