+++ src/design/plone/contenttypes/interfaces/bando.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-from plone.app.textfield import RichText
+from collective.volto.blocksfield.field import BlocksField
 from plone.autoform import directives
 from plone.supermodel import model
 from redturtle.bandi import bandiMessageFactory as _
@@ -18,91 +18,77 @@ from plone.autoform import directives as form
 class IBandoAgidSchema(IBandoSchema):
     """ A Dexterity schema for Annoucements """
 
-    form.order_after(riferimenti_bando="IRichTextBehavior.text")
-    riferimenti_bando = RichText(
-        title=_(
-            "riferimenti_bando_agid_label", default=u"Ulteriori informazioni"
-        ),
-        description=_(
-            "riferimenti_bando_agid_help",
-            default=u"Ulteriori informazioni non previste negli altri campi;"
-            " si può trattare di contatti o note informative la cui conoscenza"
-            " è indispensabile per la partecipazione al bando",
-        ),
+    # ridefinito, così usiamo il campo dei blocchi
+    text = BlocksField(
+        title=_("text_label", default="Testo"),
+        description=_("text_help", default="",),
         required=False,
     )
-
-    form.order_after(chiusura_procedimento_bando="IRichTextBehavior.text")
-    chiusura_procedimento_bando = schema.Date(
-        title=_rtbando(
-            "chiusura_procedimento_bando_label",
-            default=u"Closing date procedure",
-        ),
-        description=_rtbando("chiusura_procedimento_bando_help", default=u""),
-        required=False,
+    tipologia_bando = schema.Choice(
+        title=_rtbando("tipologia_bando_label", default="Announcement type"),
+        description=_rtbando("tipologia_bando_help", default=""),
+        vocabulary="redturtle.bandi.tipologia.vocabulary",
+        required=True,
     )
-
-    form.order_after(scadenza_bando="IRichTextBehavior.text")
-    scadenza_bando = schema.Datetime(
-        title=_rtbando(
-            "scadenza_bando_label", default=u"Expiration date and time"
-        ),
-        description=_rtbando(
-            "scadenza_bando_help",
-            default=u"Deadline to participate in the announcement",
-        ),
-        required=False,
-    )
-
-    form.order_after(scadenza_domande_bando="IRichTextBehavior.text")
-    scadenza_domande_bando = schema.Datetime(
-        title=_(
-            "scadenza_domande_bando_label",
-            default=u"Termine per le richieste di chiarimenti",
-        ),
-        description=_(
-            "scadenza_domande_bando_help",
-            default=u"Data entro la quale sarà possibile far pervenire domande"
-            " e richieste di chiarimento a chi eroga il bando",
-        ),
-        required=False,
-    )
-
-    form.order_after(ente_bando="IRichTextBehavior.text")
-    directives.widget(
-        "ente_bando",
-        AjaxSelectFieldWidget,
-        vocabulary="redturtle.bandi.enti.vocabulary",
-    )
-    ente_bando = schema.Tuple(
-        title=_rtbando(u"ente_label", default=u"Authority"),
-        description=_rtbando(
-            u"ente_help", default=u"Select some authorities."
-        ),
-        required=False,
-        defaultFactory=getDefaultEnte,
-        value_type=schema.TextLine(),
-        missing_value=None,
-    )
-
-    form.order_after(destinatari="IRichTextBehavior.text")
-    directives.widget(destinatari=CheckBoxFieldWidget)
     destinatari = schema.List(
-        title=_rtbando("destinatari_label", default=u"Recipients"),
+        title=_rtbando("destinatari_label", default="Recipients"),
         description=_rtbando("destinatari_help", default=""),
         required=True,
         value_type=schema.Choice(
             vocabulary="redturtle.bandi.destinatari.vocabulary"
         ),
     )
+    ente_bando = schema.Tuple(
+        title=_rtbando("ente_label", default="Authority"),
+        description=_rtbando("ente_help", default="Select some authorities."),
+        required=False,
+        defaultFactory=getDefaultEnte,
+        value_type=schema.TextLine(),
+        missing_value=None,
+    )
+    scadenza_domande_bando = schema.Datetime(
+        title=_(
+            "scadenza_domande_bando_label",
+            default="Termine per le richieste di chiarimenti",
+        ),
+        description=_(
+            "scadenza_domande_bando_help",
+            default="Data entro la quale sarà possibile far pervenire domande"
+            " e richieste di chiarimento a chi eroga il bando",
+        ),
+        required=False,
+    )
+    scadenza_bando = schema.Datetime(
+        title=_rtbando(
+            "scadenza_bando_label", default="Expiration date and time"
+        ),
+        description=_rtbando(
+            "scadenza_bando_help",
+            default="Deadline to participate in the announcement",
+        ),
+        required=False,
+    )
 
-    form.order_after(tipologia_bando="IRichTextBehavior.text")
-    directives.widget(tipologia_bando=RadioFieldWidget)
-    tipologia_bando = schema.Choice(
-        title=_rtbando("tipologia_bando_label", default=u"Announcement type"),
-        description=_rtbando("tipologia_bando_help", default=""),
-        vocabulary="redturtle.bandi.tipologia.vocabulary",
-        required=True,
+    chiusura_procedimento_bando = schema.Date(
+        title=_rtbando(
+            "chiusura_procedimento_bando_label",
+            default="Closing date procedure",
+        ),
+        description=_rtbando("chiusura_procedimento_bando_help", default=""),
+        required=False,
+    )
+
+    riferimenti_bando = BlocksField(
+        title=_(
+            "riferimenti_bando_agid_label", default="Ulteriori informazioni"
+        ),
+        description=_(
+            "riferimenti_bando_agid_help",
+            default="Ulteriori informazioni non previste negli altri campi;"
+            " si può trattare di contatti o note informative la cui conoscenza"
+            " è indispensabile per la partecipazione al bando",
+        ),
+        required=False,
     )
 
     ufficio_responsabile = RelationList(
@@ -117,7 +103,7 @@ class IBandoAgidSchema(IBandoSchema):
         required=False,
         default=[],
         value_type=RelationChoice(
-            title=_(u"Ufficio responsabile"),
+            title=_("Ufficio responsabile"),
             vocabulary="plone.app.vocabularies.Catalog",
         ),
     )
@@ -135,10 +121,18 @@ class IBandoAgidSchema(IBandoSchema):
         required=False,
         default=[],
         value_type=RelationChoice(
-            title=_(u"Area"), vocabulary="plone.app.vocabularies.Catalog"
+            title=_("Area"), vocabulary="plone.app.vocabularies.Catalog"
         ),
     )
 
+    # widgets
+    directives.widget(
+        "ente_bando",
+        AjaxSelectFieldWidget,
+        vocabulary="redturtle.bandi.enti.vocabulary",
+    )
+    directives.widget(destinatari=CheckBoxFieldWidget)
+    directives.widget(tipologia_bando=RadioFieldWidget)
     form.widget(
         "ufficio_responsabile",
         RelatedItemsFieldWidget,