+++ src/design/plone/contenttypes/tests/test_ct_documento.py
@@ -31,6 +31,7 @@ class TestDocument(unittest.TestCase):
                 "design.plone.contenttypes.behavior.argomenti_documento",
                 "design.plone.contenttypes.behavior.descrizione_estesa_documento",
                 "design.plone.contenttypes.behavior.additional_help_infos",
+                "collective.dexteritytextindexer",
             ),
         )