+++ src/design/plone/contenttypes/tests/test_ct_event.py
@@ -47,7 +47,7 @@ class TestEvent(unittest.TestCase):
                 "plone.versioning",
                 "plone.locking",
                 "plone.constraintypes",
-                "volto.blocks",
+                "collective.dexteritytextindexer",
             ),
         )