+++ src/design/plone/contenttypes/tests/test_ct_pagina_argomento.py
@@ -32,6 +32,5 @@ class TestPaginaArgomento(unittest.TestCase):
                 "plone.leadimage",
                 "collective.dexteritytextindexer",
                 "volto.blocks",
-                "design.plone.contenttypes.behavior.additional_help_infos",
             ),
         )