+++ src/design/plone/contenttypes/tests/test_ct_servizio.py
@@ -11,7 +11,6 @@ from plone.app.testing import setRoles
 from plone.app.testing import SITE_OWNER_NAME
 from plone.app.testing import SITE_OWNER_PASSWORD
 from plone.app.testing import TEST_USER_ID
-from plone.app.textfield.value import RichTextValue
 from plone.restapi.testing import RelativeSession
 
 import unittest
@@ -127,12 +126,9 @@ class TestServizioApi(unittest.TestCase):
             container=self.portal,
             type="Servizio",
             title="Test servizio",
-            a_chi_si_rivolge=RichTextValue(
-                raw="destinatari",
-                mimeType="text/html",
-                outputMimeType="text/html",
-                encoding="utf-8",
-            ),
+            a_chi_si_rivolge={
+                "blocks": {"foo": {"searchableText": "destinatari"}}
+            },
         )
 
         res = api.content.find(SearchableText="destinatari")
@@ -147,12 +143,7 @@ class TestServizioApi(unittest.TestCase):
             container=self.portal,
             type="Servizio",
             title="Test servizio",
-            chi_puo_presentare=RichTextValue(
-                raw="chi",
-                mimeType="text/html",
-                outputMimeType="text/html",
-                encoding="utf-8",
-            ),
+            chi_puo_presentare={"blocks": {"foo": {"searchableText": "chi"}}},
         )
 
         res = api.content.find(SearchableText="chi")
@@ -167,12 +158,7 @@ class TestServizioApi(unittest.TestCase):
             container=self.portal,
             type="Servizio",
             title="Test servizio",
-            come_si_fa=RichTextValue(
-                raw="come",
-                mimeType="text/html",
-                outputMimeType="text/html",
-                encoding="utf-8",
-            ),
+            come_si_fa={"blocks": {"foo": {"searchableText": "come"}}},
         )
 
         res = api.content.find(SearchableText="come")
@@ -187,12 +173,9 @@ class TestServizioApi(unittest.TestCase):
             container=self.portal,
             type="Servizio",
             title="Test servizio",
-            cosa_si_ottiene=RichTextValue(
-                raw="ottenere",
-                mimeType="text/html",
-                outputMimeType="text/html",
-                encoding="utf-8",
-            ),
+            cosa_si_ottiene={
+                "blocks": {"foo": {"searchableText": "ottenere"}}
+            },
         )
 
         res = api.content.find(SearchableText="ottenere")
@@ -207,12 +190,7 @@ class TestServizioApi(unittest.TestCase):
             container=self.portal,
             type="Servizio",
             title="Test servizio",
-            cosa_serve=RichTextValue(
-                raw="serve",
-                mimeType="text/html",
-                outputMimeType="text/html",
-                encoding="utf-8",
-            ),
+            cosa_serve={"blocks": {"foo": {"searchableText": "serve"}}},
         )
 
         res = api.content.find(SearchableText="serve")
@@ -227,12 +205,9 @@ class TestServizioApi(unittest.TestCase):
             container=self.portal,
             type="Servizio",
             title="Test servizio",
-            ulteriori_informazioni=RichTextValue(
-                raw="aiuto",
-                mimeType="text/html",
-                outputMimeType="text/html",
-                encoding="utf-8",
-            ),
+            ulteriori_informazioni={
+                "blocks": {"123456": {"searchableText": "aiuto"}}
+            },
         )
 
         res = api.content.find(SearchableText="aiuto")